Moved all scaffold composables to scaffold directory.
