+++ android/app/src/main/java/com/koleff/kare_android/ui/compose/scaffolds/MainScreenScaffold.kt
@@ -1,4 +1,4 @@
-package com.koleff.kare_android.ui.compose
+package com.koleff.kare_android.ui.compose.scaffolds
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
