+++ android/app/src/main/java/com/koleff/kare_android/ui/compose/screen/SettingsScreen.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
-import com.koleff.kare_android.ui.compose.MainScreenScaffold
+import com.koleff.kare_android.ui.compose.scaffolds.MainScreenScaffold
 import com.koleff.kare_android.ui.compose.SettingsList
 
 @Composable
